+++ docs/pages/components/message/index.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
-import { Message, Button, ButtonToolbar, SelectPicker, toaster } from 'rsuite';
+import { Message, Button, ButtonToolbar, SelectPicker, useToaster } from 'rsuite';
 import DefaultPage from '@/components/Page';
 
 export default function Page() {
-  return <DefaultPage dependencies={{ Message, Button, ButtonToolbar, SelectPicker, toaster }} />;
+  return (
+    <DefaultPage dependencies={{ Message, Button, ButtonToolbar, SelectPicker, useToaster }} />
+  );
 }