+++ src/Form/test/Form.test.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Form, { FormInstance } from '../Form';
+
+const formRef = React.createRef<FormInstance>();
+
+<Form ref={formRef}></Form>;
+
+formRef.current?.check();