+++ src/toaster/index.tsx
@@ -1,3 +1,5 @@
 import toaster from './toaster';
+
 export type { Toaster } from './toaster';
+export { default as useToaster } from './useToaster';
 export default toaster;