+++ src/utils/constants.ts
@@ -1,4 +1,5 @@
 export const SIZE = ['lg', 'md', 'sm', 'xs'];
+export const COLUMN_SIZE = ['xxl', 'xl', 'lg', 'md', 'sm', 'xs'];
 export const STATUS = ['success', 'warning', 'error', 'info'];
 export const COLOR = ['red', 'orange', 'yellow', 'green', 'cyan', 'blue', 'violet'];