+++ stylelint.config.js
@@ -1,53 +1,81 @@
 module.exports = {
   extends: ['stylelint-config-standard', 'stylelint-config-prettier'],
+  customSyntax: 'postcss-less',
   rules: {
-    // Overwrite stylelint-config-standard rules
-    // ----------------------------------
+    // IE 11 does not seem to support percentage format
+    'alpha-value-notation': 'number',
 
-    // Don't check if there is a whitespace after the function.
-    // Style-lint can't distinguish less mixin or css function. Disable this check item directly here.
-    'function-whitespace-after': null,
-    // Require a leading zero for fractional numbers less than 1.
-    'number-leading-zero': 'always',
-
-    // Extended rules
-    // - The most restrictive vendor-prefix limit is enabled, which is determined by the data of {@link https://caniuse.com}
-    // ----------------------------------
+    // fix less detached ruleset
+    'at-rule-no-unknown': null,
 
     // Forbidden the at rule (eg: @keyframes ) to use the browser prefix.
     'at-rule-no-vendor-prefix': true,
-    // fix less detached ruleset
-    'at-rule-no-unknown': null,
+
+    // IE 11 does not support modern css notation
+    // https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgb#browser_compatibility
+    'color-function-notation': 'legacy',
+
+    'declaration-colon-newline-after': null,
+
     // Specify an allowed-list of attributes and units allowed in the declaration.
     'declaration-property-unit-allowed-list': {
       'font-size': ['px', 'em', '%'],
       '/^animation/': ['s']
     },
-    // Forbidden @media  use the browser prefix.
-    'media-feature-name-no-vendor-prefix': true,
+
+    // Allow using Less-specific functions
+    'function-no-unknown': [
+      true,
+      {
+        ignoreFunctions: ['fade', 'floor', 'lighten']
+      }
+    ],
+
+    // Don't check if there is a whitespace after the function.
+    // Style-lint can't distinguish less mixin or css function. Disable this check item directly here.
+    'function-whitespace-after': null,
+
     // Limit the depth of nesting allowed.
     'max-nesting-depth': 5,
+
+    // Forbidden @media  use the browser prefix.
+    'media-feature-name-no-vendor-prefix': true,
+
     // Don't forbidden low specificity selectors from appearing after high specificity selectors.
     // [Specificity description]{@link https://www.w3.org/TR/css3-selectors/#specificity}
     'no-descending-specificity': null,
+
+    // Allow placing @import at-rule in style blocks (Less-specific syntax)
+    'no-invalid-position-at-import-rule': null,
+
+    // Require a leading zero for fractional numbers less than 1.
+    'number-leading-zero': 'always',
+
+    // Forbidden use the browser prefix.
+    'property-no-vendor-prefix': true,
+
     // Class selector naming rules use kebab-case. eg: .class-name
     // FIXME: this regex covers LESS parametric mixin definitions
     //        which was supposed to be ignored from this rule by stylelint
     //        @see https://github.com/stylelint/stylelint/issues/5258
     'selector-class-pattern': '^[a-z0-9]+(-?[a-z0-9]+)*(\\(.*\\))?$',
+
     // ID selector naming rules use camel case , excepted `icon-`. eg: #loginButton,#icon-login
     'selector-id-pattern': '^([a-z]+([A-Z][a-z]+)?)||(icon-([a-z-]+))$',
-    // Forbidden use the browser prefix.
-    'property-no-vendor-prefix': true,
-    'selector-no-vendor-prefix': true,
-    'value-no-vendor-prefix': true,
+
     // Limit the number of composite selectors in a selector.
     // TODO consider narrowing this limit
     'selector-max-compound-selectors': 4,
+
     // Limit the number of ID selectors in a selector.
+    // TODO consider narrowing this limit to 0
     'selector-max-id': 1,
+
     // Limit the number of universal selectors in a selector.
     'selector-max-universal': 1,
+
+    'selector-no-vendor-prefix': true,
+
     // Specify an allowed-list of allowed pseudo-class selectors.(Only use pseudo-classes supported by IE10+ browsers)
     // [CSS selector browser support]{@link https://labs.qianduan.net/css-selector/}
     'selector-pseudo-class-allowed-list': [
@@ -81,10 +109,10 @@ module.exports = {
       'focus-within',
       'focus-visible'
     ],
+
     // String using single quotes
     'string-quotes': 'single',
-    'declaration-colon-newline-after': null,
-    'value-list-comma-newline-after': 'always-multi-line',
+
     // font-family can contain uppercase letters
     // @see https://github.com/stylelint/stylelint/issues/4622#issuecomment-594641799
     'value-keyword-case': [
@@ -92,6 +120,9 @@ module.exports = {
       {
         ignoreKeywords: ['BlinkMacSystemFont']
       }
-    ]
+    ],
+    'value-list-comma-newline-after': 'always-multi-line',
+
+    'value-no-vendor-prefix': true
   }
 };